Started main menu
Started adding in the main menu widget. Imported a free to use font for the title and button text.
Set up the blueprint for the player character so that I can run around in the test level and check CMC defaults, camera, etc.
Set up a small test level to travel to for testing of the training room.